+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Removed layout rounding no longer needed since [#64] ([#95]).
+
 ### Added
 
 - [Add support for the iPhone SE 2](https://github.com/square/Blueprint/pull/96) in `ElementPreview`.
@@ -206,6 +208,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.3.1]: https://github.com/square/Blueprint/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/square/Blueprint/compare/0.2.2...0.3.0
 [0.2.2]: https://github.com/square/Blueprint/releases/tag/0.2.2
+[#95]: https://github.com/square/Blueprint/pull/95
 [#72]: https://github.com/square/Blueprint/pull/72
 [#68]: https://github.com/square/Blueprint/pull/68
 [#67]: https://github.com/square/Blueprint/pull/67